Add docs improvement, add list of unknown questions
