+++ qa-admin/src/main/java/ru/volpi/qaadmin/dto/question/UnknownQuestionResponse.java
@@ -1,0 +1,4 @@
+package ru.volpi.qaadmin.dto.question;
+
+public record UnknownQuestionResponse(Long id, String text) {
+}
