+++ qa-admin/src/main/java/ru/volpi/qaadmin/service/UnknownQuestionService.java
@@ -1,0 +1,13 @@
+package ru.volpi.qaadmin.service;
+
+import ru.volpi.qaadmin.dto.question.Answer;
+import ru.volpi.qaadmin.dto.question.QuestionResponse;
+import ru.volpi.qaadmin.dto.question.UnknownQuestionResponse;
+
+import java.util.List;
+
+public interface UnknownQuestionService {
+    QuestionResponse addAnswer(Answer answer);
+
+    List<UnknownQuestionResponse> findAll();
+}
