+++ qa-admin/src/main/java/ru/volpi/qaadmin/service/impl/QuestionServiceImpl.java
@@ -7,17 +7,12 @@ import lombok.extern.slf4j.Slf4j;
 import org.springframework.transaction.annotation.Transactional;
 import ru.volpi.qaadmin.domain.question.Question;
 import ru.volpi.qaadmin.domain.question.Questions;
-import ru.volpi.qaadmin.dto.question.Answer;
 import ru.volpi.qaadmin.dto.question.QuestionRegistration;
 import ru.volpi.qaadmin.dto.question.QuestionResponse;
 import ru.volpi.qaadmin.dto.question.QuestionUpdate;
-import ru.volpi.qaadmin.exception.category.CategoryNotFoundException;
 import ru.volpi.qaadmin.exception.question.QuestionNotFoundException;
 import ru.volpi.qaadmin.exception.question.QuestionValidationException;
-import ru.volpi.qaadmin.repository.CategoryRepository;
 import ru.volpi.qaadmin.repository.QuestionRepository;
-import ru.volpi.qaadmin.repository.UnknownQuestionRepository;
-import ru.volpi.qaadmin.service.EmailService;
 import ru.volpi.qaadmin.service.QuestionService;
 import ru.volpi.qaadmin.service.annotation.TransactionalService;
 
@@ -30,12 +25,6 @@ import java.util.List;
 public class QuestionServiceImpl implements QuestionService {
 
     private final QuestionRepository questionRepository;
-
-    private final CategoryRepository categoryRepository;
-
-    private final UnknownQuestionRepository unknownQuestionRepository;
-
-    private final EmailService emailService;
 
     private final Validator validator;
 
@@ -97,36 +86,5 @@ public class QuestionServiceImpl implements QuestionService {
         }
         this.questionRepository.deleteById(id);
         return id;
-    }
-
-    @Transactional
-    @Override
-    public QuestionResponse addAnswer(final Answer answer) {
-        log.info("Answer came {}", answer);
-        return QuestionResponse.from(
-            this.unknownQuestionRepository.findById(answer.getUnknownQuestionId())
-                .map(
-                    question -> {
-                        final Question answered = this.questionRepository.save(
-                            Question.builder()
-                                .text(question.getText())
-                                .answer(answer.getText())
-                                .category(
-                                    this.categoryRepository.findByNameIgnoreCase(answer.getCategory())
-                                        .orElseThrow(() -> new CategoryNotFoundException(answer.getCategory()))
-                                )
-                                .build()
-                        );
-                        this.unknownQuestionRepository.deleteById(answer.getUnknownQuestionId());
-                        this.emailService.sendNotification(
-                            question.getEmail(),
-                            "Ответ на Ваш вопрос добавлен",
-                            question.getText()
-                        );
-                        log.info("Answer {} added", answer);
-                        return answered;
-                    }
-                ).orElseThrow(() -> new QuestionNotFoundException(answer.getUnknownQuestionId()))
-        );
     }
 }
