+++ qa-admin/src/test/java/ru/volpi/qaadmin/service/impl/UnknownQuestionServiceImplTest.java
@@ -1,0 +1,35 @@
+package ru.volpi.qaadmin.service.impl;
+
+import org.junit.jupiter.api.DisplayName;
+import org.junit.jupiter.api.Test;
+import org.springframework.beans.factory.annotation.Autowired;
+import ru.volpi.qaadmin.TestcontainersTest;
+import ru.volpi.qaadmin.dto.question.Answer;
+import ru.volpi.qaadmin.dto.question.QuestionResponse;
+import ru.volpi.qaadmin.repository.UnknownQuestionRepository;
+
+import static org.assertj.core.api.Assertions.assertThat;
+
+class UnknownQuestionServiceImplTest extends TestcontainersTest {
+
+    @Autowired
+    private UnknownQuestionServiceImpl unknownService;
+
+    @Autowired
+    private UnknownQuestionRepository unknownQuestionRepository;
+
+    @Test
+    @DisplayName("Adds answer to unknown question")
+    void addsAnswer() {
+        final String answerPhrase = "Ответ на ваш вопрос";
+        final String secondCategory = "Вторая категория";
+        final QuestionResponse actual = this.unknownService.addAnswer(
+            new Answer(111L, answerPhrase, secondCategory)
+        );
+        assertThat(this.unknownQuestionRepository.findById(111L).isPresent()).isFalse();
+        assertThat(actual.answer()).isEqualTo(answerPhrase);
+        assertThat(actual.text()).isEqualTo("Мой новый вопрос");
+        assertThat(actual.categoryName()).isEqualTo(secondCategory);
+    }
+
+}
